Revive INSTANT_MILLIS instead of NANO_OF_INSTANT_SECONDS to represent %Q
In addition, many tests are imported from:
https://github.com/embulk/embulk/blob/v0.9.17/embulk-core/src/test/java/org/embulk/spi/time/TestTimestampParser.java

----

A bug in processing %Q (milliseconds since epoch) motivated this change.
%Q was not working fine for negative numbers before this fix. It was because
of a mistake in division and remainder of negative integers.

The mistake was caused by complicated code from unnecessary separation of
millisecond. Then, this commit reverts the change in #32, and fixes the
original problem in a different way.

The original problem of #32 was unnecessary limited MAX_VALUE and MIN_VALUE
of INSTANT_SECONDS. #32 solved it by separating %Q into an integer second pard
and a fractional part. Instead, this commit solves it by maintaining both
INSTANT_SECONDS and INSTANT_MILLIS at the same time for %Q.

For %s, INSTANT_SECONDS is set, and INSTANT_MILLIS is cleared. Then, %s does
not have the unnecessary limitation of MAX_VALUE and MIN_VALUE.

For %Q, both INSTANT_SECONDS and INSTANT_MILLIS are set. Then, %Q has the
limitation of MAX_VALUE and MIN_VALUE, but the limitation is unavoidable as
the millisecond is originally represented in long.

Finally, this representation does not damage the capability of INSTANT_SECONDS,
and then just extends for milliseconds since epoch.