Apply a time offset in a correct manner
It has been another division bug than #39. It used "/" and "%" operators when
applying a time offset (e.g. +09:00), but the dividend could be negative. As
a result, it failed to parse a timestamp like "2019-05-03 03:00:00 +09:00".

The time offset "+09:00" should make the example to "2019-05-02 18:00:00Z"
with carrying up to its day. But because of the bug above, it went to
"2019-05-03 (-06):00:00Z", and it failed.

The original Ruby code uses "Numeric#divmod" for this calculation.
https://github.com/ruby/ruby/blob/v2_6_3/lib/time.rb#L174
https://docs.ruby-lang.org/en/2.6.0/Numeric.html#method-i-divmod

In Java, it should have been "Math.floorDiv" and "Math.floorMod", instead of
the "/" and "%" operators.